+++ src/main/java/com/umust/dobonglife/domain/user/domain/repository/custom/UserRepositoryCustom.java
@@ -1,5 +1,5 @@
 package com.umust.dobonglife.domain.user.domain.repository.custom;
 
 public interface UserRepositoryCustom {
-    long decreaseBalance(Long userId, int amount);
+    long decreaseBalance(Long userId, Long amount);
 }
